+++ CHANGELOG.md
@@ -11,4 +11,18 @@
 
 ## 0.2.0
 
-* Use custom criteria.
+* Use custom criteria. Use string keys to indicate custom criteria, e.g. `made_purchase`.
+* Make cool down intervals nullable.
+* More tests.
+
+## 0.2.1
+
+* Update README.
+
+## 0.2.2
+
+* Fix changelog formatting.
+
+## 0.2.3
+
+* Update README.